LayoutTests/fast/js/kde/evil-n-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL (new Error()).message should be undefined. Was Unknown error
PASS ''.split(/.*/).length is 0
FAIL window.successfullyParsed should be undefined. Was true

TEST COMPLETE
